Делаю рейсы.Нужно сделать валидацию и временную таблицу

+++ main/java/com/svop/service/handbooks/NomerReysService.java
@@ -2,6 +2,7 @@ package com.svop.service.handbooks;
 
 import com.svop.View.NomerReysView;
 import com.svop.View.NomerReysViewRequest;
+import com.svop.tables.Handbooks.Aircompany;
 import com.svop.tables.Handbooks.NomerReys;
 import com.svop.tables.Handbooks.NomerReysRepository;
 import org.slf4j.Logger;
@@ -13,29 +14,57 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class NomerReysService {
     private static Logger logger= LoggerFactory.getLogger(NomerReysService.class.getName());
     @Autowired
-    NomerReysRepository nomerReysRepository;
+    private NomerReysRepository nomerReysRepository;
     @Autowired
-    AircompaniesService aircompaniesService;
+    private AircompaniesService aircompaniesService;
     public List<NomerReysView> getNomerReysFromAircompany(Integer id){
+
         logger.info("Запрос на получение номеров рейсов авиакомпании");
+        if (id==null)
+        {
+            logger.error("Нулевое значение авиакомпании!");
+        }
         List<NomerReys> nomerReys=nomerReysRepository.findByAircompany_Id(id);
         List<NomerReysView> nomerReysViews=new ArrayList<>();
         for(NomerReys item:nomerReys) nomerReysViews.add(new NomerReysView(item));
         return nomerReysViews;
     }
+
+    //----------------------------------------------------------------------------
+    public Optional<NomerReys> getByid(Integer id)
+    {
+        if (id==null)
+        {
+            logger.error("Ошибка. Запрос с пустым id");
+            return null;
+        }
+        logger.info("Запрос на получение номера реса с id ="+id);
+        return nomerReysRepository.findById(id);
+
+    }
     public void delete(Iterable<Integer> id_list)
     {
+        logger.info("Запрос на удаление списка номеров рейсов");
+        if (id_list==null) {
+            logger.error("Ошибка удаления номеров рейсов");
+            return;
+        }
+        logger.info("Удаление номеров рейсов завершено");
         nomerReysRepository.deleteByIdIn(id_list);
     }
+    ////////-----------------------------------------------------------------
+
     public ResponseEntity<String> save(List<NomerReysView> nomerReysViews)
     {
        return new ResponseEntity<>("Успех",HttpStatus.OK);
     }
+    //Сохранить объект списка номеров
     public ResponseEntity<String> save(NomerReysViewRequest nomerReysViewRequest)
     {
         logger.info("Запрос на сохранение номеров рейсов авиакомпании");
@@ -44,7 +73,7 @@ public class NomerReysService {
             logger.error("Авиакомпания не опеределена");
             return new ResponseEntity<String>("Авиакомпания не опеределена",HttpStatus.BAD_REQUEST);
         }
-      //  List<NomerReys> nomerReys=new ArrayList<>();
+
         for(NomerReysView nomerReysView:nomerReysViewRequest.getNomers())
         {
             NomerReys nomer=new NomerReys();
@@ -58,4 +87,6 @@ public class NomerReysService {
         //nomerReysRepository.save(nomerReys);
        return new ResponseEntity<String>("Успех",HttpStatus.OK);
     }
+
+
 }
